+++ jme3-core/src/main/java/com/jme3/opencl/MemoryAccess.java
@@ -32,12 +32,21 @@
 package com.jme3.opencl;
 
 /**
- * Specifies how a buffer object can be accessed
+ * Specifies how a buffer object can be accessed by the kernel.
  * @author Sebastian Weiss
  * @see Buffer
  */
 public enum MemoryAccess {
+    /**
+     * A kernel can both read and write the buffer.
+     */
 	READ_WRITE,
+    /**
+     * A kernel can only write this buffer.
+     */
 	WRITE_ONLY,
+    /**
+     * A kernel can only read this buffer
+     */
 	READ_ONLY
 }
