+++ jme3-core/src/main/java/com/jme3/opencl/OpenCLException.java
@@ -32,10 +32,14 @@
 package com.jme3.opencl;
 
 /**
- *
+ * Generic OpenCL exception, can be thrown in every method of this package.
+ * The error code and its name is reported in the message string as well as the OpenCL call that
+ * causes this exception. Please refer to the official OpenCL specification
+ * to see what might cause this exception.
  * @author Sebastian Weiss
  */
 public class OpenCLException extends RuntimeException {
+    private static final long serialVersionUID = 8471229972153694848L;
 
 	private final int errorCode;
 	
@@ -63,6 +67,9 @@ public class OpenCLException extends RuntimeException {
 		this.errorCode = errorCode;
 	}
 
+    /**
+     * @return the error code
+     */
 	public int getErrorCode() {
 		return errorCode;
 	}
